fix: update rules "Last Updated" date
Commit 92623649a90aa84b66b33931f7eff6623256864f created rule 5.7, but
apparently forgot to change the "Last Updated" date. This changes the
date to Feb 25th (which is the UTC time; the true last update date was
Feb 24th for many users due to timezones). In the future this might need
to either be automatically updated, removed, or rules-updaters will need
to start remembering to change this date.